+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Globally disable workspace support via makefile config #264
 * Support additional rust script runners #216
 * Use task cwd in condition and installation scripts #259 (**backward compatability break**)
+* Check descriptor min_version before deserializing #268 (thanks @roblabla)
 * Profile name passed incorrectly via forked sub tasks #263
 
 ### v0.21.0 (2019-06-26)